# api.yaml
swagger: "2.0.0"
info:
  title: Test API Documentation
  description: Api Documentation for Test App Cyber Flow
  version: 1.0.0
  contact:
    email: choudharysidharth082000@gmail.com
  servers:
    - url: "http://localhost:3000"
      description: "Local Server Running"
    - url: "https://testcyberflow.herokuapp.com"
      description: "Main Server"
schemes: [http, https]
paths:     
...
  /v1/analytics/getCountSubCategory/{categoryID}:
    get:
      tags:
        - name: Analytics
      summary: Getting count of SUb Categories of a particular id
      description: lorem
      parameters:
        - name: categoryID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
  /v1/analytics/getAllReferredStudents/{teacherID}:
    get:
      tags:
        - name: Analytics
      summary: Get Referrred Students
      description: lorem
      parameters:
        - name: teacherID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
  /v1/analytics/getStudentByDate/{teacherID}:
    post:
      tags:
        - name: Analytics
      summary: Referred Student on date
      description: lorem
      parameters:
        - name: teacherID
          in: path
          type: string
          required: true
        - name: bodyData
          in: body
          schema:
            type: object
            properties:
              date:
                type: string
      responses:
        200:
          description: OK
  /v1/analytics/getReferredStudentsDate/{teacherID}:
    post:
      tags:
        - name: Analytics
      summary: Referred Student on date
      description: lorem
      parameters:
        - name: teacherID
          in: path
          type: string
          required: true
        - name: bodyData
          in: body
          schema:
            type: object
            properties:
              date:
                type: string
      responses:
        200:
          description: OK
...
  /v1/analytics/getStudentsAtDate/{offset}/{limit}:
    post:
      tags:
        - name: Analytics
      summary: Getting all students at a particular Date
      description: lorem
      parameters:
        - name: offset 
          in: path
          type: string
          required: true
        - name: limit 
          in: path
          type: string
          required: true
        - in: body
          name: dates
          schema:
            type: object
            properties:
              from:
                type: number
              end:
                type: number
      responses:
        200:
          description: OK
  /v1/analytics/getAllPayout/{teacherID}:
    get:
      tags:
        - name: Analytics
      summary: Getting All checkoutsof teacher
      description: lorem
      parameters:
        - name: teacherID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
...
  /v1/analytics/getUsersLastWeek/{offset}/{limit}:
    get:
      tags:
        - name: Analytics
      summary: Users Joined Last Week
      description: Getting users joined last week
      parameters:
        - name: offset
          in: path
          type: string
          required: true  
        - name: limit
          in: path
          type: string
          required: true  
      responses:
        200:
          description: OK
...
  /v1/analytics/getUsersLastMonth/{offset}/{limit}:
    get:
      tags:
        - name: Analytics
      summary: Getting Users last month
      description: Getting all the users joined last month
      parameters:
        - name: offset
          in: path
          type: string
          required: true
        - name: limit
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK  
...
  /v1/analytics/studentAnalytics/{offset}/{limit}:
    get:
      tags:
        - name: Analytics
      summary: Getting Students Joined Today 
      description: lorem
      parameters:
        - name: offset
          in: path
          type: string
          required: true
        - name: limit
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
  /v1/analytics/teacherAnalytics/{offset}/{limit}:
    get:
      tags:
        - name: Analytics
      summary: Getting Students Joined Today 
      description: lorem
      parameters:
        - name: offset
          in: path
          type: string
          required: true
        - name: limit
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
  /v1/analytics/changeMainContentHindi/{testID}:
    put:
      tags:
        - name: Additional
      summary: Shift To MainContent Hindi
      description: lorem
      parameters:
        - name: testID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
  # /v1/analytics/getReferredStudentsDate/{teacherID}:
  #   get:
  #     tags:
  #       - name: Analytics
  #     summary: Getting Referred students
  #     description: OK
  #     parameters:
  #       - name: teacherID
  #         in: path
  #         type: string
  #         required: true
  #       - name: referral
  #         in: body
  #         schema:
  #           type: object
  #           properties:
  #             date:
  #               type: string
  #     responses:
  #       200:
  #         description: OK
...
  /v1/analytics/searchItems:
    post:
      tags:
        - name: Search Items
      summary: Search Items
      description: lorem
      consumes:
        multipart/form-data
      produces:
        application/json
      parameters:
        - name: search
          in: body
          schema:
            type: object
            properties:
              search:
                type: string          
      responses:
        200:
          description: OK
...